+++ src/constants/index.ts
@@ -5,6 +5,7 @@ export * from "./statuses";
 export * from "./phases";
 export * from "./game";
 export * from "./avatar-scripts";
+export * from "./navi";
 export * from "./host";
 export * from "./realtime";
 export * from "./voice";